replace each controller to each file and made reexport
