+++ controllers/contactControllers/getAllContacts.js
@@ -1,0 +1,8 @@
+import { Contact } from "../../models/index.js";
+
+export const getAllContacts = async (req, res) => {
+  const contacts = await Contact.find();
+  res.json(contacts);
+};
+
+export default getAllContacts;
